nee/requirements.txt
fastapi==0.108.0
mysql-connector-python==8.2.0
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
uvicorn==0.25.0
python-dotenv~=1.0.1
opencv-python==4.5.5.64
pyzbar==0.1.8
requests==2.31.0

pillow
keras
opencv-python
numpy~=1.26.4
keras~=3.3.3
pillow~=10.3.0
qrcode~=7.4.2
